requirements.txt
requests
python-dotenv
openai
ipython
pytz